--- a/ethereum/core/src/jmh/java/org/hyperledger/besu/ethereum/vm/operations/BenchmarkHelper.java
+++ b/ethereum/core/src/jmh/java/org/hyperledger/besu/ethereum/vm/operations/BenchmarkHelper.java
@@ -16,7 +16,6 @@ package org.hyperledger.besu.ethereum.vm.operations;
 
 import static org.mockito.Mockito.mock;
 
-import org.apache.tuweni.units.bigints.UInt256;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Hash;
 import org.hyperledger.besu.datatypes.Wei;
@@ -29,52 +28,51 @@ import java.util.Random;
 
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
+import org.apache.tuweni.units.bigints.UInt256;
 
 public class BenchmarkHelper {
   public static MessageFrame createMessageCallFrame() {
     return MessageFrame.builder()
-            .worldUpdater(mock(WorldUpdater.class))
-            .originator(Address.ZERO)
-            .gasPrice(Wei.ONE)
-            .blobGasPrice(Wei.ONE)
-            .blockValues(mock(BlockValues.class))
-            .miningBeneficiary(Address.ZERO)
-            .blockHashLookup((__, ___) -> Hash.ZERO)
-            .type(MessageFrame.Type.MESSAGE_CALL)
-            .initialGas(Long.MAX_VALUE)
-            .address(Address.ZERO)
-            .contract(Address.ZERO)
-            .inputData(Bytes32.ZERO)
-            .sender(Address.ZERO)
-            .value(Wei.ZERO)
-            .apparentValue(Wei.ZERO)
-            .code(CodeV0.EMPTY_CODE)
-            .completer(__ -> {
-            })
-            .build();
+        .worldUpdater(mock(WorldUpdater.class))
+        .originator(Address.ZERO)
+        .gasPrice(Wei.ONE)
+        .blobGasPrice(Wei.ONE)
+        .blockValues(mock(BlockValues.class))
+        .miningBeneficiary(Address.ZERO)
+        .blockHashLookup((__, ___) -> Hash.ZERO)
+        .type(MessageFrame.Type.MESSAGE_CALL)
+        .initialGas(Long.MAX_VALUE)
+        .address(Address.ZERO)
+        .contract(Address.ZERO)
+        .inputData(Bytes32.ZERO)
+        .sender(Address.ZERO)
+        .value(Wei.ZERO)
+        .apparentValue(Wei.ZERO)
+        .code(CodeV0.EMPTY_CODE)
+        .completer(__ -> {})
+        .build();
   }
 
   public static MessageFrame createMessageCallFrameWithCallData(final Bytes callData) {
     return MessageFrame.builder()
-            .worldUpdater(mock(WorldUpdater.class))
-            .originator(Address.ZERO)
-            .gasPrice(Wei.ONE)
-            .blobGasPrice(Wei.ONE)
-            .blockValues(mock(BlockValues.class))
-            .miningBeneficiary(Address.ZERO)
-            .blockHashLookup((__, ___) -> Hash.ZERO)
-            .type(MessageFrame.Type.MESSAGE_CALL)
-            .initialGas(Long.MAX_VALUE)
-            .address(Address.ZERO)
-            .contract(Address.ZERO)
-            .inputData(callData)
-            .sender(Address.ZERO)
-            .value(Wei.ZERO)
-            .apparentValue(Wei.ZERO)
-            .code(CodeV0.EMPTY_CODE)
-            .completer(__ -> {
-            })
-            .build();
+        .worldUpdater(mock(WorldUpdater.class))
+        .originator(Address.ZERO)
+        .gasPrice(Wei.ONE)
+        .blobGasPrice(Wei.ONE)
+        .blockValues(mock(BlockValues.class))
+        .miningBeneficiary(Address.ZERO)
+        .blockHashLookup((__, ___) -> Hash.ZERO)
+        .type(MessageFrame.Type.MESSAGE_CALL)
+        .initialGas(Long.MAX_VALUE)
+        .address(Address.ZERO)
+        .contract(Address.ZERO)
+        .inputData(callData)
+        .sender(Address.ZERO)
+        .value(Wei.ZERO)
+        .apparentValue(Wei.ZERO)
+        .code(CodeV0.EMPTY_CODE)
+        .completer(__ -> {})
+        .build();
   }
 
   /**
@@ -103,7 +101,12 @@ public class BenchmarkHelper {
     return Bytes.wrap(data);
   }
 
-  static void fillPoolsForCallData(final Bytes[] sizePool, final Bytes[] destOffsetPool, final Bytes[] srcOffsetPool, final int dataSize, final boolean fixedSrcDst) {
+  static void fillPoolsForCallData(
+      final Bytes[] sizePool,
+      final Bytes[] destOffsetPool,
+      final Bytes[] srcOffsetPool,
+      final int dataSize,
+      final boolean fixedSrcDst) {
     for (int i = 0; i < sizePool.length; i++) {
       sizePool[i] = Bytes.wrap(UInt256.valueOf(dataSize));