--- a/evm/src/main/java/org/hyperledger/besu/evm/operation/CallDataCopyOperation.java
+++ b/evm/src/main/java/org/hyperledger/besu/evm/operation/CallDataCopyOperation.java
@@ -38,16 +38,19 @@ public class CallDataCopyOperation extends AbstractOperation {
   @Override
   public OperationResult execute(final MessageFrame frame, final EVM evm) {
     final long memOffset = clampedToLong(frame.popStackItem());
-    final long sourceOffset = clampedToLong(frame.popStackItem());
+    final Bytes sourceOffsetBytes = frame.popStackItem();
     final long numBytes = clampedToLong(frame.popStackItem());
-
     final long cost = gasCalculator().dataCopyOperationGasCost(frame, memOffset, numBytes);
     if (frame.getRemainingGas() < cost) {
       return new OperationResult(cost, ExceptionalHaltReason.INSUFFICIENT_GAS);
     }
 
-    final Bytes callData = frame.getInputData();
+    if (numBytes == 0) {
+      return new OperationResult(cost, null); // Skip writeMemory call
+    }
 
+    final Bytes callData = frame.getInputData();
+    final long sourceOffset = clampedToLong(sourceOffsetBytes);
     frame.writeMemory(memOffset, sourceOffset, numBytes, callData, true);
 
     return new OperationResult(cost, null);